requirements.txt
torch
torchvision
numpy
scikit-learn
biopython
tensorboard
pyyaml